git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-repro-enh-20120205@9592 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ rutil/dns/DnsStub.hxx
@@ -37,7 +37,7 @@ class GetDnsCacheDumpHandler
    public:
       GetDnsCacheDumpHandler() {}
       virtual ~GetDnsCacheDumpHandler() {}
-      virtual void onDnsCacheDumpRetrieved(std::pair<unsigned int, unsigned int> key, const Data& dnsCache) = 0;
+      virtual void onDnsCacheDumpRetrieved(std::pair<unsigned long, unsigned long> key, const Data& dnsCache) = 0;
 };
 
 template<typename T>
